+++ layer-api/src/main/java/org/layer/admin/retrospect/controller/dto/AdminRetrospectsGetResponse.java
@@ -1,0 +1,16 @@
+package org.layer.admin.retrospect.controller.dto;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.layer.domain.retrospect.dto.AdminRetrospectGetResponse;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Getter
+public class AdminRetrospectsGetResponse {
+
+	private final List<AdminRetrospectGetResponse> retrospects;
+	private final Integer totalCount;
+
+}
