+++ layer-api/src/main/java/org/layer/admin/space/controller/dto/AdminSpaceCountGetResponse.java
@@ -1,0 +1,9 @@
+package org.layer.admin.space.controller.dto;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(name = "AdminSpaceCountGetResponse", description = "우리 팀이 만든 스페이스를 제외하고 기간 내에 만들어진 스페이스 개수를 리턴합니다.")
+public record AdminSpaceCountGetResponse(
+        @Schema(description = "기간 내에 만들어진 스페이스 개수", example = "11")
+        Long spaceCount) {
+}
