+++ layer-api/src/main/resources/application-dev.yml
@@ -18,7 +18,7 @@ spring:
 
   data:
     redis:
-      host: ${DEV_REDIS_HOST}
+      host: localhost
       port: ${DEV_REDIS_PORT}
       password: ${DEV_REDIS_PASSWORD}
 
@@ -83,6 +83,7 @@ discord:
     space-url: ${DISCORD_SPACE_URL}
     member-url: ${DISCORD_MEMBER_URL}
     error-url: ${DISCORD_ERROR_URL}
+    member-activity-url: ${DISCORD_MEMBER_ACTIVITY_URL}
 
 # actuator
 management:
