+++ layer-api/src/main/resources/application-local.yml
@@ -88,3 +88,4 @@ discord:
     space-url: ${DISCORD_SPACE_URL}
     member-url: ${DISCORD_MEMBER_URL}
     error-url: ${DISCORD_ERROR_URL}
+    member-activity-url: ${DISCORD_MEMBER_ACTIVITY_URL}
