+++ layer-external/src/main/java/org/layer/discord/event/MemberActivityEvent.java
@@ -1,0 +1,11 @@
+package org.layer.discord.event;
+
+import java.util.Map;
+
+public record MemberActivityEvent(
+	Map<Long, Map<String, Integer>> activities
+) {
+	public static MemberActivityEvent of(Map<Long, Map<String, Integer>> activities){
+		return new MemberActivityEvent(activities);
+	}
+}
